+++ src/TomorrowDAOServer.Application.Contracts/Common/Security/SecretProvider.cs
@@ -1,9 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using AElf.Types;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Newtonsoft.Json;
 using TomorrowDAOServer.Common.Cache;
 using TomorrowDAOServer.Common.Dtos;
 using TomorrowDAOServer.Common.HttpClient;
@@ -16,12 +18,13 @@ namespace TomorrowDAOServer.Common.Security;
 public interface ISecretProvider
 {
     Task<string> GetSecretWithCacheAsync(string key);
+    Task<string> GetSignatureFromHashAsync(string publicKey, Hash hash);
 }
 
 public class SecretProvider : ISecretProvider, ITransientDependency
 {
     private const string GetSecurityUri = "/api/app/thirdPart/secret";
-
+    private const string GetSignatureUri = "/api/app/signature";
     private readonly ILogger<SecretProvider> _logger;
     private readonly IOptionsMonitor<SecurityServerOptions> _securityServerOption;
     private readonly ILocalMemoryCache<string> _secretCache;
@@ -49,6 +52,31 @@ public class SecretProvider : ISecretProvider, ITransientDependency
                 DateTimeOffset.Now.AddSeconds(_securityServerOption.CurrentValue.SecretCacheSeconds)
         });
     }
+    
+    public async Task<string> GetSignatureFromHashAsync(string publicKey, Hash hash)
+    {
+        try
+        {
+            var signatureSend = new SendSignatureDto
+            {
+                PublicKey = publicKey,
+                HexMsg = hash.ToHex(),
+            };
+
+            var url = Uri(GetSignatureUri);
+            var resp = await _httpProvider.InvokeAsync<CommonResponseDto<SignResponseDto>>(HttpMethod.Post,
+                url, body: JsonConvert.SerializeObject(signatureSend), header: SecurityServerHeader());
+            AssertHelper.IsTrue(resp?.Success ?? false, "Signature response failed");
+            AssertHelper.NotEmpty(resp!.Data?.Signature, "Signature response empty");
+            return resp.Data!.Signature;
+        }
+        catch (Exception e)
+        {
+            _logger.LogError(e, "CallSignatureServiceFailed, err: {err}, hash: {body}", e.ToString(),
+                JsonConvert.SerializeObject(hash.ToHex()));
+            return null;
+        }
+    }
 
     private async Task<string> GetSecretAsync(string key)
     {
@@ -64,7 +92,7 @@ public class SecretProvider : ISecretProvider, ITransientDependency
         return EncryptionHelper.DecryptFromHex(resp.Data, _securityServerOption.CurrentValue.AppSecret);
     }
 
-    public Dictionary<string, string> SecurityServerHeader(params string[] signValues)
+    private Dictionary<string, string> SecurityServerHeader(params string[] signValues)
     {
         var signString = string.Join(CommonConstant.EmptyString, signValues);
         return new Dictionary<string, string>
@@ -72,5 +100,16 @@ public class SecretProvider : ISecretProvider, ITransientDependency
             ["appid"] = _securityServerOption.CurrentValue.AppId,
             ["signature"] = EncryptionHelper.EncryptHex(signString, _securityServerOption.CurrentValue.AppSecret)
         };
+    }
+    
+    private class SendSignatureDto
+    {
+        public string PublicKey { get; set; }
+        public string HexMsg { get; set; }
+    }
+
+    private class SignResponseDto
+    {
+        public string Signature { get; set; }
     }
 }
